requirements.txt
# Python dependencies (FastAPI, SQLAlchemy, etc.). 
# Dependencies for this service only  

fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
pydantic[email]
python-jose[cryptography]
pytest
# PyJWT
# jwt
pydantic-settings
python-multipart
alembic
azure.communication.email
azure.storage.blob
azure.ai.formrecognizer